Multihoming: combined commit for review
servers: Remove synchronization for HeartbeatThread#shutdown().

The synchronization is a leftover from a previous stage of development.
In recent commits synchronization has been removed from the other
methods and can be removed from shutdown too.
Synchronization is not necessary since shutdown is idempotent.

servers: Add the matching network in CIDR notation for local addresses
to the address mappings.

servers: Use POSIX signal USR2 on OSDs to renew their address mappings.

servers: Mark open questions regarding the addressMappings.

servers: Remove org.apache.commons.net dependency.

servers: Try finding the localHostAddress and set it as globally
accessible.
Prefer local networks over global ones when resolving an UUID.

servers: Wait on a monitor in HeartbeatThread to enable immediate
address renewals.

servers: Move the address renewal signal handler to the HeartbeatThread
and make it configurable.

This makes the HeartbeatThread responsible for the signal handling. OSD
and MRC initialize it in regard to the new setting 'renewal_signal'.
The new setting is briefly documented.

servers: Add config paramater for multihoming and address renewal signals.

servers: Adapt endpoint registering to the introduced properties.

With this commit there are different possibilities to determine the endpoints
that will be propagated at the DIR.

1. multihomed.enabled is set to true and hostname is empty:
   Every valid address will be used with the corresponding network.
   It will be attempted to find the local host automatically and then use it as
   the default.

2. multihomed.enabled is set to true and hostname is set:
   As in the first case, but the configured hostname will be used to determine
   the default.

3. multihomes.enabled is not set and neither hostname nor listen.address
   It will be attempted to find the first non-local address and use it as a
   single endpoint matching any network. If this isn't successful it will be
   attempted to use the first local address. If this fails, too, the localhost
   will be used.

4. multihome.enabled is not set but hostname or listen.address
   The configured hostname or if not set listen.address will be used to as a
   single endpoint matching any network.